Move Category composables from Composables.kt to CategoryComposables.kt
